feat!: simplify template file handling and path processing
BREAKING CHANGE: Handle template file types as a single `files` list in
template manifest and use a single `path` property to handle both
source and target paths in project file generation. This change expects
a leading '_' in template path to indicate a hidden target
file/directory, and '.ftl' ending to identify a FreeMarker file for
preprocessing.